Revert "pmt-ptw"
This reverts commit b15c4e70ba5e1c36e4faaa80edc196f0d620a3a4.

+++ eclipse-collections/src/main/java/org/eclipse/collections/impl/bag/immutable/ImmutableArrayBag.java
@@ -84,7 +84,7 @@ public class ImmutableArrayBag<T>
             public void value(T each, int count)
             {
                 newKeys[this.index] = each;
-                newCounts[this.index] = ((Integer)(count)).intValue();
+                newCounts[this.index] = count;
                 this.index++;
             }
         });
