+++ eclipse-collections/src/main/java/org/eclipse/collections/impl/block/procedure/ChainedProcedure.java
@@ -43,7 +43,7 @@ public final class ChainedProcedure<T> implements Procedure<T>
         int size = this.procedures.size();
         for (int i = 0; i < size; i++)
         {
-            this.procedures.get(((Integer)(i)).intValue()).value(object);
+            this.procedures.get(i).value(object);
         }
     }
