+++ eclipse-collections/src/main/java/org/eclipse/collections/impl/lazy/parallel/list/ListIterableBatch.java
@@ -38,7 +38,7 @@ public class ListIterableBatch<T> extends AbstractBatch<T> implements RootListBa
     {
         for (int i = this.chunkStartIndex; i < this.chunkEndIndex; i++)
         {
-            procedure.value(this.list.get(((Integer)(i)).intValue()));
+            procedure.value(this.list.get(i));
         }
     }
