+++ eclipse-collections/src/main/java/org/eclipse/collections/impl/map/mutable/UnifiedMap.java
@@ -318,7 +318,7 @@ public class UnifiedMap<K, V> extends AbstractMutableMap<K, V>
         // This function ensures that hashCodes that differ only by
         // constant multiples at each bit position have a bounded
         // number of collisions (approximately 8 at default load factor).
-        int h = key == null ? 0 : ((Integer)(key.hashCode())).intValue();
+        int h = key == null ? 0 : key.hashCode();
         h ^= h >>> 20 ^ h >>> 12;
         h ^= h >>> 7 ^ h >>> 4;
         return (h & (this.table.length >> 1) - 1) << 1;
